+++ src/main/java/com/obervatorio_pedagogico/backend/presentation/dto/dashboard/response/ConjuntoDadosResponse.java
@@ -14,5 +14,5 @@ import lombok.Setter;
 public class ConjuntoDadosResponse {
     private String legenda;
     
-    private List<Integer> dados;
+    private List<Float> dados;
 }
